[Spaces] Move getSpaceIdFromPath to a separate package (#221485)
## Summary

This PR creates a separate package with `getSpaceIdFromPath` exported to
avoid creating a circular dependency in
https://github.com/elastic/kibana/pull/220138

---------

(cherry picked from commit 7ac80fa5a6352c7ede93691e60d3e4da81e051d0)

# Conflicts:
#	.github/CODEOWNERS
#	package.json
#	tsconfig.base.json
#	x-pack/platform/plugins/shared/spaces/tsconfig.json
#	yarn.lock
