+++ .github/CODEOWNERS
@@ -729,6 +729,74 @@ src/platform/packages/shared/kbn-profiling-utils @elastic/obs-ux-infra_services-
 x-pack/platform/packages/private/kbn-random-sampling @elastic/kibana-visualizations
 src/platform/packages/shared/kbn-react-field @elastic/kibana-data-discovery
 src/platform/packages/shared/kbn-react-hooks @elastic/obs-ux-logs-team
+src/platform/packages/shared/kbn-recently-accessed @elastic/appex-sharedux
+src/platform/packages/shared/kbn-repo-info @elastic/kibana-operations
+src/platform/packages/shared/kbn-resizable-layout @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-rison @elastic/kibana-operations
+src/platform/packages/shared/kbn-router-to-openapispec @elastic/kibana-core
+src/platform/packages/shared/kbn-router-utils @elastic/obs-ux-logs-team
+src/platform/packages/shared/kbn-rrule @elastic/response-ops
+src/platform/packages/shared/kbn-rule-data-utils @elastic/security-detections-response @elastic/response-ops @elastic/obs-ux-management-team
+src/platform/packages/shared/kbn-safer-lodash-set @elastic/kibana-security
+src/platform/packages/shared/kbn-saved-search-component @elastic/obs-ux-logs-team
+src/platform/packages/shared/kbn-scout @elastic/appex-qa
+src/platform/packages/shared/kbn-search-api-panels @elastic/search-kibana
+src/platform/packages/shared/kbn-search-connectors @elastic/search-kibana
+src/platform/packages/shared/kbn-search-errors @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-search-response-warnings @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-search-types @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-security-hardening @elastic/kibana-security
+src/platform/packages/shared/kbn-securitysolution-ecs @elastic/security-threat-hunting-explore
+src/platform/packages/shared/kbn-securitysolution-es-utils @elastic/security-detection-engine
+src/platform/packages/shared/kbn-securitysolution-io-ts-types @elastic/security-detection-engine
+src/platform/packages/shared/kbn-securitysolution-io-ts-utils @elastic/security-detection-engine
+src/platform/packages/shared/kbn-securitysolution-rules @elastic/security-detection-engine
+src/platform/packages/shared/kbn-server-http-tools @elastic/kibana-core
+src/platform/packages/shared/kbn-server-route-repository @elastic/obs-knowledge-team
+src/platform/packages/shared/kbn-server-route-repository-client @elastic/obs-knowledge-team
+src/platform/packages/shared/kbn-server-route-repository-utils @elastic/obs-knowledge-team
+src/platform/packages/shared/kbn-shared-svg @elastic/obs-ux-infra_services-team
+src/platform/packages/shared/kbn-shared-ux-utility @elastic/appex-sharedux
+src/platform/packages/shared/kbn-sort-predicates @elastic/kibana-visualizations
+src/platform/packages/shared/kbn-spaces-utils @elastic/kibana-security
+src/platform/packages/shared/kbn-sse-utils @elastic/obs-knowledge-team
+src/platform/packages/shared/kbn-sse-utils-client @elastic/obs-knowledge-team
+src/platform/packages/shared/kbn-sse-utils-server @elastic/obs-knowledge-team
+src/platform/packages/shared/kbn-std @elastic/kibana-core
+src/platform/packages/shared/kbn-storage-adapter @elastic/observability-ui
+src/platform/packages/shared/kbn-storybook @elastic/kibana-operations
+src/platform/packages/shared/kbn-telemetry @elastic/kibana-core @elastic/obs-ai-assistant
+src/platform/packages/shared/kbn-telemetry-config @elastic/kibana-core
+src/platform/packages/shared/kbn-test @elastic/kibana-operations @elastic/appex-qa
+src/platform/packages/shared/kbn-test-jest-helpers @elastic/kibana-operations @elastic/appex-qa
+src/platform/packages/shared/kbn-test-subj-selector @elastic/kibana-operations @elastic/appex-qa
+src/platform/packages/shared/kbn-timerange @elastic/obs-ux-logs-team
+src/platform/packages/shared/kbn-tooling-log @elastic/kibana-operations
+src/platform/packages/shared/kbn-traced-es-client @elastic/observability-ui
+src/platform/packages/shared/kbn-tracing @elastic/kibana-core @elastic/obs-ai-assistant
+src/platform/packages/shared/kbn-triggers-actions-ui-types @elastic/response-ops
+src/platform/packages/shared/kbn-try-in-console @elastic/search-kibana
+src/platform/packages/shared/kbn-typed-react-router-config @elastic/obs-knowledge-team @elastic/obs-ux-infra_services-team
+src/platform/packages/shared/kbn-ui-actions-browser @elastic/appex-sharedux
+src/platform/packages/shared/kbn-ui-theme @elastic/kibana-operations
+src/platform/packages/shared/kbn-unified-data-table @elastic/kibana-data-discovery @elastic/security-threat-hunting-investigations
+src/platform/packages/shared/kbn-unified-doc-viewer @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-unified-field-list @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-unified-histogram @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-unified-tabs @elastic/kibana-data-discovery
+src/platform/packages/shared/kbn-unsaved-changes-prompt @elastic/kibana-management
+src/platform/packages/shared/kbn-use-tracked-promise @elastic/obs-ux-logs-team
+src/platform/packages/shared/kbn-user-profile-components @elastic/kibana-security
+src/platform/packages/shared/kbn-utility-types @elastic/kibana-core
+src/platform/packages/shared/kbn-utility-types-jest @elastic/kibana-operations
+src/platform/packages/shared/kbn-utils @elastic/kibana-operations
+src/platform/packages/shared/kbn-visualization-ui-components @elastic/kibana-visualizations
+src/platform/packages/shared/kbn-visualization-utils @elastic/kibana-visualizations
+src/platform/packages/shared/kbn-xstate-utils @elastic/obs-ux-logs-team
+src/platform/packages/shared/kbn-zod @elastic/kibana-core
+src/platform/packages/shared/kbn-zod-helpers @elastic/security-detection-rule-management
+src/platform/packages/shared/presentation/presentation_containers @elastic/kibana-presentation
+src/platform/packages/shared/presentation/presentation_publishing @elastic/kibana-presentation
 src/platform/packages/shared/react/kibana_context/common @elastic/appex-sharedux
 src/platform/packages/shared/react/kibana_context/render @elastic/appex-sharedux
 src/platform/packages/shared/react/kibana_context/root @elastic/appex-sharedux
