+++ src/platform/packages/shared/kbn-spaces-utils/README.md
@@ -1,0 +1,3 @@
+# @kbn/spaces-utils
+
+Package containing utils related to Spaces.
