+++ src/platform/packages/shared/kbn-spaces-utils/kibana.jsonc
@@ -1,0 +1,9 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/spaces-utils",
+  "owner": [
+    "@elastic/kibana-security"
+  ],
+  "group": "platform",
+  "visibility": "shared"
+}
