+++ src/platform/packages/shared/kbn-spaces-utils/src/common/spaces_url_parser.ts
@@ -1,11 +1,13 @@
 /*
  * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { DEFAULT_SPACE_ID } from '../constants';
+import { DEFAULT_SPACE_ID } from './constants';
 
 const spaceContextRegex = /^\/s\/([a-z0-9_\-]+)/;
