+++ tsconfig.base.json
@@ -1898,6 +1898,8 @@
       "@kbn/spaces-plugin/*": ["x-pack/platform/plugins/shared/spaces/*"],
       "@kbn/spaces-test-plugin": ["x-pack/test/spaces_api_integration/common/plugins/spaces_test_plugin"],
       "@kbn/spaces-test-plugin/*": ["x-pack/test/spaces_api_integration/common/plugins/spaces_test_plugin/*"],
+      "@kbn/spaces-utils": ["src/platform/packages/shared/kbn-spaces-utils"],
+      "@kbn/spaces-utils/*": ["src/platform/packages/shared/kbn-spaces-utils/*"],
       "@kbn/spec-to-console": ["packages/kbn-spec-to-console"],
       "@kbn/spec-to-console/*": ["packages/kbn-spec-to-console/*"],
       "@kbn/sse-utils": ["src/platform/packages/shared/kbn-sse-utils"],
